Remove public namespace for controllers and views
