--- app/controllers/challenges/application_controller.rb
@@ -1,5 +1,0 @@
-module Challenges
-  class ApplicationController < ApplicationController
-    scoped_authorization :challenges
-  end
-end
