--- app/controllers/challenges/public/application_controller.rb
@@ -1,8 +1,0 @@
-module Challenges
-  module Public
-    class ApplicationController < ApplicationController
-      allow_unauthenticated_access
-      scoped_authorization :challenges, :public
-    end
-  end
-end
