+++ app/controllers/challenges/submissions/application_controller.rb
@@ -1,0 +1,7 @@
+module Challenges
+  module Submissions
+    class ApplicationController < ::ApplicationController
+      scoped_authorization :challenges
+    end
+  end
+end
