+++ app/views/challenges/models/_model.html.erb
