+++ app/views/challenges/models/_scores.html.erb
