+++ app/views/challenges/models/index.html.erb
