+++ app/views/challenges/models/show.html.erb
