+++ app/views/challenges/tasks/_menu.html.erb
