+++ app/views/challenges/tasks/_task.html.erb
