+++ app/views/challenges/tasks/leaderboards/_filters.html.erb
