+++ app/views/challenges/tasks/leaderboards/_leaderboard.html.erb
