+++ app/views/challenges/test_sets/_menu.html.erb
