+++ app/views/challenges/test_sets/_show.html.erb
