+++ app/views/challenges/test_sets/leaderboards/_leaderboard.html.erb
