+++ app/views/challenges/test_sets/show.html.erb
@@ -3,5 +3,5 @@
 <% end %>
 <% breadcrumb :test_set, @test_set %>
 <%= mltop_title @test_set.name %>
-<%= render "challenges/public/test_sets/menu", test_set: @test_set %>
+<%= render "challenges/test_sets/menu", test_set: @test_set %>
 <%= render "show", test_set: @test_set, task_test_sets: @task_test_sets %>
